Taking input images, added stb_image library.
Added three image types: jpg, jfif and the png converted jfif. Looked around and saw a few libraries but noticed two videos on stb_image being good for image processing so I'm running with them. Not too sure if I need to focus too much on CMake at the moment but may review later for testing/running. Have to use a wierd mix of MSBuild with cmake to have it run correctly. Tutorial said install is the way so I can run it as a normal function like ls but can't seem to get it to work just yet. Still setting things up.