Added depthFirstSearch
Added depth first search, it doesn't quite get the shortest path. Primarily it focuses on a single path until the path fails, after which it continues with a prior path usually one step backwards but in a different direction. Does show an interesting view of the full traversal of the successful path.