Added aStarAdjacent
Added a rough take on the A* algorithm. Using a priority queue. The current version needs a bit of hammering out. It performs well and looks to access less spaces but the runtime is abnormally long, especially as it should be performing far less operations compared to the other algorithms.